Added code for plot3 and fixed a bug in plot2
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   plot2/belief_prop/csv_gen.py
	modified:   plot2/fuzzy/csv_gen.py
	modified:   plot2/intersect/csv_gen.py
	new file:   plot3/belief_prop/.gitignore
	new file:   plot3/belief_prop/csv_gen.py
	new file:   plot3/fuzzy/.gitignore
	new file:   plot3/fuzzy/csv_gen.py
	new file:   plot3/intersect/.gitignore
	new file:   plot3/intersect/csv_gen.py

+++ plot2/intersect/csv_gen.py
@@ -96,7 +96,7 @@ code_rates = [1/2, 2/3, 3/4, 5/6, 7/8]
 jobs = mp.Queue() # Create a job queue for workers
 for snr, snr_db in zip(snr_linear, snr_db):
     for code_rate in code_rates:
-        num_message_bits = n * code_rate
+        num_message_bits = int(n * code_rate)
         jobs.put((snr, snr_db, code_rate, num_message_bits))
 
 # Create a manager for the shared list
